refactor: moved function to load tasks from JSON configuration into func.c
