+++ include/func.h
@@ -10,5 +10,6 @@ void run_scheduler_concurrent(void);
 void handle_shutdown(int signum);
 void setup_signal_handlers();
 void start_watchdog();
+void load_tasks_from_json(const char *filename);
 
 #endif // FUNC_H
